GoLive 20170405

+++ clean_boot.sh
@@ -21,10 +21,12 @@ main(){
 	local l_starttime=$(date +%s)
 	f_init_from_ini
 	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ Cleanup after boot"
-	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ Start DB integrity check"
-	local l_db_consistent=$(sqlite3 $i_db "pragma integrity_check;")
-	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ DB integrity check done"
-	rm $_base/lock
+	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ Start DB integrity check for $g_database"
+	local l_db_consistent=$(sqlite3 $g_database "pragma integrity_check;")
+	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ DB integrity check on $g_database done"
+	if [[ -e "$_base/lock" ]]; then
+		rm $_base/lock
+	fi
 	log_always "$_scriptlocal $LINENO ++++BOOT++++++++++++++++++ runtime script $(( $(date +%s) - $l_starttime )) secs - End of processing"
 }
